Create db migrations
Create migrations for movies, auditoria, showtimes and purchases tables. This is my first time testing activerecord's pedantry on the level of auditorium/auditoria so this should be interesting. Also I forgot that I uninstalled everything I had installed with homebrew the other week when homebrew and yarn got in a fight so that was a fun time with postgresql just now.